+++ tests/integration/assetmanagement/test_asset_management.py
@@ -54,6 +54,9 @@ def client(enterprise_config: HttpConfiguration) -> AssetManagementClient:
 @pytest.mark.integration
 @pytest.mark.enterprise
 class TestAssetManagement:
+    _workspace = "2300760d-38c4-48a1-9acb-800260812337"
+    """Used the main-test default workspace since the client
+    for creating a workspace has not been added yet"""
 
     _create_assets_request = [
         CreateAssetRequest(
@@ -77,7 +80,7 @@ class TestAssetManagement:
                 date="2022-06-07T18:58:05.000Z",
             ),
             is_NI_asset=True,
-            workspace="2300760d-38c4-48a1-9acb-800260812337",
+            workspace=_workspace,
             location=AssetLocationForCreate(
                 state=AssetPresence(asset_presence=AssetPresenceStatus.PRESENT)
             ),
@@ -119,6 +122,25 @@ class TestAssetManagement:
         assert len(assets) == 1
         asset_id = assets[0].id
         assert asset_id is not None
+
+    def test__link_files__link_succeded(
+        self, client: AssetManagementClient, create_asset
+    ):
+        self._create_assets_request[0].model_number = 101
+        create_response = create_asset(self._create_assets_request)
+
+        asset_id = (
+            create_response.assets[0].id
+            if create_response.assets and create_response.assets[0].id
+            else None
+        )
+
+        assert asset_id is not None
+
+        file_ids = ["608a5684800e325b48837c2a"]
+        link_files_response = client.link_files(asset_id=asset_id, file_ids=file_ids)
+
+        assert link_files_response is None
 
     def test__delete_asset__returns_deleted_asset(self, client: AssetManagementClient):
         self._create_assets_request[0].model_number = 102