+++ tests/integration/systems/test_systems.py
@@ -27,7 +27,8 @@ def client(enterprise_config: HttpConfiguration) -> SystemsClient:
 
 
 workspace_id = "2300760d-38c4-48a1-9acb-800260812337"
-"""Constant represent id of the workspace."""
+"""Used the main-test default workspace since the client
+for creating a workspace has not been added yet"""
 
 
 @pytest.fixture